move login/logout login into AuthProvider; remove auth controller
